requirements.txt
asgiref==3.5.2
asttokens==2.0.8
autopep8==1.7.0
backcall==0.2.0
colorama==0.4.5
debugpy==1.6.2
decorator==5.1.1
Django==4.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-decouple==2.1
djangorestframework==3.14.0
entrypoints==0.4
executing==0.10.0
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
pywin32==304
pyzmq==23.2.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
toml==0.10.2
tornado==6.2
traitlets==5.3.0
tzdata==2022.2
wcwidth==0.2.5
